chore: move combined/separate state components
