+++ src/__tests__/components/pages/race/dashboard/state-combined.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 import { useStaticQuery } from 'gatsby'
-import { StateCombined } from '~components/pages/race/dashboard/state-combined'
+import { StateCombined } from '~components/pages/race/state-combined'
 
 beforeEach(() => {
   useStaticQuery.mockImplementation(() => ({
