+++ src/components/pages/race/state.module.scss
